Relocate DRO_PVVP_Gallery_Factory class to includes/gallery/factories for improved structure
